enterprise-ai-engine/requirements.txt
# ResumeGuard — Full Project Requirements

# Dashboard
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.20.0

# Data Validation
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0

# LLM / AI
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
anthropic>=0.28.0
openai>=1.30.0

# PDF Parsing
pymupdf>=1.24.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0